Some fixes in the magma gemms. These have to be tested, especially the dgemm because I removed something that Pend had and there may be possibilities for some type of segmentation faults on the GPU. I haven't seen problems so far. Performance is very good and uniform. Peng, good job in generalizing the gemms.
Peng, in what cases did you see problems if we remove "(bll<k)*"?
I forgot if I already asked this. I tested without it and everything works,
directl testing and through LU.

If there are no problems removing it I may take it out from sgemm.
We loose performance with it when everything is divisible (about 10GFlop/s in
double and 20 in single). Otherwise we have to provide a second implementation
for nice dimensions.

By the way, does anybody know what is the size of a matrix that we can bind
to a texture in Fermi? NVIDIA people are doing some test on this. I tested for
large problems without checking and didn't see a problem. Is this possible to
be an issues only for the cards with 6GB.


git-svn-id: https://icl.cs.utk.edu/svn/magma/branches/sc_release@343 02bb5f2a-90ca-11de-bb24-8aedcc835bc8