This version of zlabrd.cpp comes from revision 348 with some minor tweaks to make it compile. This version works for single and double precision real. Both complex versions are still broken. And this is despite of all the conjugation calls inserted (ZLACG) and all Transposes converted to Conjugate transposes.
I've tried a number of things over the last few days. I still don't know that is wrong.
The only reason why I'm checking it in so that we have something that works better that
revision 370. Ideally, we'd like the revision 370 (because it's clean) and that works.
I'm not there yet.



git-svn-id: https://icl.cs.utk.edu/svn/magma/branches/sc_release@480 02bb5f2a-90ca-11de-bb24-8aedcc835bc8