Now this is available in complex as well. The only missing part was a cast of double to single triangular matrix. But there is another problem
./testing_dsposv_gpu -N 3009

gives me nan. This can be due to translations as I am testing on tesla and this
was working in previous releases. Mathieu, any idea where the problem may be?

Interesting enough, the complex version works for this N.
Going through the cast that I added though in real (which should be the
only difference), still doesn't work in real.



git-svn-id: https://icl.cs.utk.edu/svn/magma/branches/sc_release@740 02bb5f2a-90ca-11de-bb24-8aedcc835bc8