Here is the zlat2c. I don't see why there was a need for such complicated transfer. Rajib, if you read, why did we need all that shared memory allocation? I removed it as single precision is allocated anyway and we don't have to be carefull not to overwrite the lower, correspondingly upper, parts of the matrix, as this is just a working copy.
About the bug. I see I added defintion of cublasDsymv on tesla - apparantly
is missing. This makes me remember that our difference with complex is that
in complex we use CUBLAS, i.e., it could be a problem in magmablas_dsymv.
Tim, I hope you have checked these for any sides, not just multiple of the
block size. Somebody has to check if the error comes from the use of magmablas.


git-svn-id: https://icl.cs.utk.edu/svn/magma/branches/sc_release@741 02bb5f2a-90ca-11de-bb24-8aedcc835bc8