Added PAPI SDEs to GMRES / PGMRES.
Note: Initial residual wasn't set properly
because it was calculated in IF statement that was
never executed.
if (solver_par->numiter == 0) never evaluated to TRUE
because solver_par->numiter was increased prior to
IF statement.
For now, changed to: if (solver_par->numiter == 1).
Alternatively, solver_par->numiter could be increased
after the IF statement.

--HG--
branch : PAPI_SDE