hipMAGMA is finally building (and a super-basic example can be ran which uses `magma_sgemm`).
Right now, I am dealing with unimplemented vendor functions as `#define hipblasXXX mher`, and `mher` is a macro which will print out an warning message that it isn't defined. This is a work in progress, and I plan to go back and fix what I can, and use our magmablas implementations of some of them. But, for a prototype, it does 'work' for some functions.

Other than that, most of the changes are focused around changing `#ifdef HAVE_CUBLAS` to `#if defined(HAVE_CUBLAS) || defined(HAVE_HIP)` and the like. Some places, like in `control/get_n*.cpp` , we will need to find optimal values for AMD devices, but this is something we will do and publish a paper on (as Stan has suggested).

--HG--
branch : hipMAGMA